+++ numba/tests/test_print.py
@@ -94,7 +94,7 @@ class TestPrint(TestCase):
         cr = compile_isolated(pyfunc, (arraytype,), flags=enable_pyobj_flags)
         cfunc = cr.entry_point
         with captured_stdout():
-            cfunc(np.arange(10))
+            cfunc(np.arange(10, dtype=np.int32))
             self.assertEqual(sys.stdout.getvalue(),
                              '[0 1 2 3 4 5 6 7 8 9]\n')
